requirements.txt
hikari==2.0.0.dev101
hikari-lightbulb==1.2.1
edgedb==0.17.3
cattrs==1.8.0
PyYaml